+++ mex/extractors/pipeline/base.py
@@ -54,7 +54,7 @@ def create_monitor_jobs_sensor(extractor_group_names: list[str]) -> SensorDefini
         latest_start_time_extractors = last_run_time_publisher
         completed = True
 
-        actual_running = instance.get_runs(
+        jobs_are_running = instance.get_runs(
             filters=RunsFilter(
                 statuses=[
                     DagsterRunStatus.STARTING,
@@ -65,9 +65,9 @@ def create_monitor_jobs_sensor(extractor_group_names: list[str]) -> SensorDefini
                 ],
             )
         )
-        if actual_running:
+        if jobs_are_running:
             return SkipReason(
-                "No publishing because extractors are running at the moment."
+                "No publishing because other jobs are running at the moment."
             )
 
         for group in extractor_group_names: