+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Dagster sensor to run Publisher after other extractor jobs
+
 ### Changes
 
 - split seq-repo load function into new asset